use inherited storage view for CRUD in NamespaceStore
