+++ command/commands.go
@@ -38,7 +38,6 @@ import (
 
 	physPostgresql "github.com/openbao/openbao/physical/postgresql"
 	physRaft "github.com/openbao/openbao/physical/raft"
-	physXFile "github.com/openbao/openbao/physical/xfile"
 	physFile "github.com/openbao/openbao/sdk/v2/physical/file"
 	physInmem "github.com/openbao/openbao/sdk/v2/physical/inmem"
 
@@ -148,7 +147,6 @@ var (
 
 	physicalBackends = map[string]physical.Factory{
 		"file":       physFile.NewFileBackend,
-		"xfile":      physXFile.NewFileBackend,
 		"inmem_ha":   physInmem.NewInmemHA,
 		"inmem":      physInmem.NewInmem,
 		"raft":       physRaft.NewRaftBackend,
