--- command/namespace/xfile/token.json
@@ -1,8 +1,0 @@
-{
-  "policies": ["web", "stage"],
-  "meta": {
-    "user": "armon"
-  },
-  "ttl": "1h",
-  "renewable": true
-}
